Use weak linking for backward compatibility
When a symbol in a framework is defined as weakly linked, the symbol
does not have to be present at runtime for a process to continue
running. This allows taking advantage of new features while still
supporting older versions of the framework.

To build a file system that still supports osxfuse 2.x define the
following preprocessor macro before importing the framework.

GM_VERSION_MIN_REQUIRED=GM_OSXFUSE_2

Make sure to check for the existence of weakly linked symbols before
using them, otherwise the file system might crash.